requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
numpy==1.25.2
torch==2.1.1+cu121
torchvision
torchaudio
albumentations
scipy
pandas
pytorch_lightning
einops
timm
fvcore
pandas
mmengine==0.10.1
mmcv==2.1.0
opencv-python-headless
ftfy
regex
mmdet==3.3.0
mmsegmentation==1.2.2
mmpretrain==1.2.0
hydra-core
botocore
boto3
albumentations
boto3
pyarrow
pytest
ruff
isort
ruff
mypy
